updated colors and layouts
