+++ app/static/js/package.json
@@ -19,6 +19,7 @@
     "@material-ui/icons": "^1.1.0",
     "d3": "^5.8.0",
     "d3-cloud": "^1.2.5",
+    "d3-scale": "^2.2.2",
     "html-webpack-plugin": "^2.22.0",
     "path": "^0.12.7",
     "react": "^16.2.0",
